adapt drawing for graphic_2d

+++ rosen/modules/2d_graphics/src/drawing/engine_adapter/skia_adapter/skia_matrix.cpp
@@ -151,7 +151,7 @@ void SkiaMatrix::MapPoints(std::vector<Point>& dst, const std::vector<Point>& sr
         return;
     }
     if (dst.size() > count) {
-        for (int i = dst.size(); i > count; --i) {
+        for (size_t i = dst.size(); i > count; --i) {
             dst.pop_back();
         }
     }
